Update auth feature

+++ backend/app/Http/Controllers/Api/V1/Permission/PermissionController.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace App\Http\Controllers\Api\V1\User;
+namespace App\Http\Controllers\Api\V1\Permission;
 
 use App\Enums\ResponseEnum;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\User\{
+use App\Http\Requests\Permission\{
     StorePermissionRequest,
     UpdatePermissionRequest
 };
-use App\Http\Resources\User\PermissionResource;
-use App\Repositories\Interfaces\User\PermissionRepositoryInterface;
-use App\Services\Interfaces\User\PermissionServiceInterface;
+use App\Http\Resources\Permission\PermissionResource;
+use App\Repositories\Interfaces\Permission\PermissionRepositoryInterface;
+use App\Services\Interfaces\Permission\PermissionServiceInterface;
 
 class PermissionController extends Controller
 {
@@ -28,6 +28,8 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        $this->authorize('modules', 'permissions.index');
+
         $response = $this->userCatalogueService->paginate();
         $statusCode = $response['status'] == 'success' ? ResponseEnum::OK : ResponseEnum::INTERNAL_SERVER_ERROR;
         return response()->json($response, $statusCode);
@@ -38,6 +40,7 @@ class PermissionController extends Controller
      */
     public function store(StorePermissionRequest $request)
     {
+        $this->authorize('modules', 'permissions.store');
         $response = $this->userCatalogueService->create();
         $statusCode = $response['status'] == 'success' ? ResponseEnum::CREATED : ResponseEnum::INTERNAL_SERVER_ERROR;
         return response()->json($response, $statusCode);
@@ -48,6 +51,7 @@ class PermissionController extends Controller
      */
     public function show(string $id)
     {
+        $this->authorize('modules', 'permissions.show');
         $userCatalogue = new PermissionResource($this->userCatalogueRepository->findById($id));
         return response()->json([
             'status' => 'success',
@@ -62,6 +66,7 @@ class PermissionController extends Controller
      */
     public function update(UpdatePermissionRequest $request, string $id)
     {
+        $this->authorize('modules', 'permissions.update');
         $response = $this->userCatalogueService->update($id);
         $statusCode = $response['status'] == 'success' ? ResponseEnum::OK : ResponseEnum::INTERNAL_SERVER_ERROR;
         return response()->json($response, $statusCode);
@@ -73,6 +78,7 @@ class PermissionController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('modules', 'permissions.destroy');
         $response = $this->userCatalogueService->destroy($id);
         $statusCode = $response['status'] == 'success' ? ResponseEnum::OK : ResponseEnum::INTERNAL_SERVER_ERROR;
         return response()->json($response, $statusCode);
